FKS-1028 Added logging and cleaning code
